requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.4.2
blinker==1.8.2
cachelib==0.13.0
cffi==1.17.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.0
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
executing==2.1.0
filelock==3.15.3
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
gunicorn==23.0.0
ipykernel==6.29.4
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
msgspec==0.18.6
mypy-extensions==1.0.0
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.8
psycopg2==2.9.9
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
pyzmq==26.0.3
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
tzdata==2024.1
virtualenv==20.26.2
wcwidth==0.2.13
Werkzeug==3.0.3
